apps/api/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.6
requests>=2.31.0
pytz>=2023.3
PyJWT>=2.6.0
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0
google-api-python-client>=2.140.0
google-auth>=2.35.0
